@@ -4,8 +4,8 @@ publish_to: 'none'
 version: 3.5.1+46
 
 environment:
-  sdk: ">=3.5.0 <4.0.0"
-  flutter: ">=3.24.0 <4.0.0"
+  sdk: ">=3.6.0 <4.0.0"
+  flutter: ">=3.27.0 <4.0.0"
 
 dependencies:
   app_links: ^6.3.2
