@@ -9,69 +9,67 @@ import 'package:flutter/material.dart';
 /// If [leadingEntity] is specified, an [EntityImage] widget will be displayed
 /// to the left of the [title] and [subtitle]. The image will be circular if
 /// [circularLeadingImage] is `true`.
-AppBar createAppBar(String title,
+PreferredSizeWidget createAppBar(BuildContext context, String title,
         {Entity? leadingEntity,
         bool circularLeadingImage = false,
         String? subtitle,
         Color? backgroundColor,
         List<Widget>? actions,
         PreferredSizeWidget? bottom}) =>
-    AppBar(
-      backgroundColor: backgroundColor,
-      centerTitle: true,
-      title: leadingEntity != null
-          ? FittedBox(
-              fit: BoxFit.fitWidth,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  EntityImage(
-                    entity: leadingEntity,
-                    width: 40,
-                    isCircular: circularLeadingImage,
-                  ),
-                  const SizedBox(width: 8),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(title),
-                      if (subtitle != null)
-                        Text(
-                          subtitle,
-                          style: const TextStyle(fontSize: 12),
-                        )
-                    ],
-                  ),
-                ],
-              ),
-            )
-          : Column(
-              children: [
-                FittedBox(
-                  fit: BoxFit.fitWidth,
-                  child: Text(title),
+    _PreferredSizeWrapper(
+      parentBuilder: (child) => Theme(
+        data: themeDataForAppBar(context, Preferences.themeColor.value),
+        child: child,
+      ),
+      child: AppBar(
+        backgroundColor: backgroundColor,
+        centerTitle: true,
+        title: leadingEntity != null
+            ? FittedBox(
+                fit: BoxFit.fitWidth,
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    EntityImage(
+                      entity: leadingEntity,
+                      width: 40,
+                      isCircular: circularLeadingImage,
+                    ),
+                    const SizedBox(width: 8),
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(title),
+                        if (subtitle != null)
+                          Text(
+                            subtitle,
+                            style: const TextStyle(fontSize: 12),
+                          )
+                      ],
+                    ),
+                  ],
                 ),
-                if (subtitle != null)
+              )
+            : Column(
+                children: [
                   FittedBox(
                     fit: BoxFit.fitWidth,
-                    child: Text(
-                      subtitle,
-                      style: const TextStyle(fontSize: 12),
-                    ),
-                  )
-              ],
-            ),
-      actions: actions,
-      bottom: bottom != null
-          ? PreferredSize(
-              preferredSize: bottom.preferredSize,
-              child: Theme(
-                data: tabBarThemeDataForAppBar(Preferences.themeColor.value),
-                child: bottom,
+                    child: Text(title),
+                  ),
+                  if (subtitle != null)
+                    FittedBox(
+                      fit: BoxFit.fitWidth,
+                      child: Text(
+                        subtitle,
+                        style: const TextStyle(fontSize: 12),
+                      ),
+                    )
+                ],
               ),
-            )
-          : null,
+        actions: actions,
+        bottom: bottom,
+      ),
     );
 
 /// A [CircularProgressIndicator] that fits nicely in [AppBar.actions].
@@ -92,4 +90,21 @@ class AppBarLoadingIndicator extends StatelessWidget {
           ),
         ),
       );
+}
+
+class _PreferredSizeWrapper extends StatelessWidget
+    implements PreferredSizeWidget {
+  final Widget Function(Widget child) parentBuilder;
+  final PreferredSizeWidget child;
+
+  const _PreferredSizeWrapper({
+    required this.parentBuilder,
+    required this.child,
+  });
+
+  @override
+  Widget build(BuildContext context) => parentBuilder(child);
+
+  @override
+  Size get preferredSize => child.preferredSize;
 }
