@@ -4,7 +4,6 @@ import 'package:finale/services/lastfm/lastfm.dart';
 import 'package:finale/services/spotify/playlist.dart';
 import 'package:finale/util/constants.dart';
 import 'package:finale/util/formatters.dart';
-import 'package:finale/util/preferences.dart';
 import 'package:finale/widgets/base/app_bar.dart';
 import 'package:finale/widgets/base/date_time_field.dart';
 import 'package:finale/widgets/base/header_list_tile.dart';
@@ -93,14 +92,13 @@ class _BatchScrobbleViewState extends State<BatchScrobbleView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: createAppBar(
+        context,
         'Scrobble',
         actions: [
           Builder(
             builder: (context) => _isLoading
                 ? const AppBarLoadingIndicator()
                 : IconButton(
-                    disabledColor: Preferences.themeColor.value.foregroundColor
-                        .withOpacity(0.39),
                     icon: const Icon(scrobbleIcon),
                     onPressed:
                         _selection.isNotEmpty ? () => _scrobble(context) : null,
