@@ -3,7 +3,6 @@ import 'package:finale/services/lastfm/common.dart';
 import 'package:finale/services/lastfm/lastfm.dart';
 import 'package:finale/services/lastfm/track.dart';
 import 'package:finale/util/constants.dart';
-import 'package:finale/util/preferences.dart';
 import 'package:finale/widgets/base/app_bar.dart';
 import 'package:finale/widgets/base/collapsible_form_view.dart';
 import 'package:finale/widgets/base/date_time_field.dart';
@@ -109,13 +108,12 @@ class _FriendScrobbleViewState extends State<FriendScrobbleView> {
   @override
   Widget build(BuildContext context) => Scaffold(
         appBar: createAppBar(
+          context,
           'Scrobble from a friend',
           actions: [
             IconButton(
               icon: const Icon(scrobbleIcon),
               onPressed: _hasItemsToScrobble ? _scrobble : null,
-              disabledColor: Preferences.themeColor.value.foregroundColor
-                  .withOpacity(0.39),
             ),
           ],
         ),
